uni02beta: base 17.1 tripleo deployment files
The patch can deploy an almost working OSP 17.1.
Unfortunately the complete configuration for cinder volume,
manila share and glance (with NFS-based file backend) requires
a few additional tripleo variables with the private configuration
(IPs, credentials), which needs to be injected later.
The support for this is currently lacking in ci-framework
but it will be added soon.