Update glance adoption scenarios
This patch updates the test suite to clean up code a bit and make sure
it is easy to understand the scenarios where Glance is adopted.
This patch moves the glance_nfs tasks in a dedicated file, and updates
the variable names to prefix with 'glance_', which is important
considering that we have other services using the same approach.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>