Enable glance in NFS scenario
This patch updates the glance template due to the missing "enabled:
true". In addition, it adds the same overrides we have in all the
other scenarios.
Because the ctlplane creates an API with type: split, we register
a new NFS GlanceAPI based on [1] and we decommission the default
one (not used in this scenario).

[1] https://github.com/openstack-k8s-operators/glance-operator/blob/main/docs/design-decisions.md#decommissioning-a-glanceapi

Signed-off-by: Francesco Pantano <fpantano@redhat.com>