ceph_migrate: rewrite a variable to avoid linting issues
This change does not change the behavior of the code, but
it works with newer ansible-lint. The version of ansible-lint was
ancient and now it needs to be updated due to some changes in ansible
itself which led to:
ModuleNotFoundError: No module named 'ansible.parsing.yaml.constructor'

The update led to a new error when parsing the definition of the pct
variable, namely:
tests/roles/ceph_migrate/tasks/ceph_validate.yaml:42:10:
  jinja[invalid][/]: Error rendering template: float division by zero

Actual code change by Francesco Pantano.
Commit message by Luigi Toscano.