Remove pcs colocation constraints from manila-share when cephnfs
This patch adds the missing tasks related to manila adoption in
case cephnfs is the backend. In this context a pcs constraints
exists between the VIP, the ceph-nfs systemd unit and manila-share.
During the adoption process we shut down manila-share (which is
adopted as part of the procedure), but we do not take any action
to the existing A/P Ganesha instance, which still needs to stay
up.
This patch adds the missing steps documented in [1].

[1] https://github.com/openstack-k8s-operators/data-plane-adoption/blob/main/docs_user/modules/proc_stopping-openstack-services.adoc

Signed-off-by: Francesco Pantano <fpantano@redhat.com>