Do not always create a manila share type
In the adoption context based on unijobs, we assume that a workload
exists in the source cloud. Therefore, we do not need to create a new
manila share type in the adopted environment, because it wouldn't be
realistic.
This patch allows to optionally perform this task, so we can still keep
it in standalone, but we don't need to execute it for unijobs.

Jira: https://issues.redhat.com/browse/OSPRH-5712

Signed-off-by: Francesco Pantano <fpantano@redhat.com>