Create retry_delay var for each place
There is no need to make so often queries if the service is ready.
It some roles, delay is 2, in some other 5 and other 6.
Let's increase delay from 2 to 5 as default value and parametrize it
to overwrite the value if needed. Such overwrite might be helpful
for hosts that does not have enough compute resources.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>