cinder adoption: add support for PureStorage/FC
Support the adoption of a cinder/PureStorage backend.
This initial implementation only supports FiberChannel, but the code is generic
enough that it should be easy to add other supported protocols, following
the pattern already followed for the cinder/NetApp adoption.

In fact this code follows a lot the workflow already used for the
cinder/NetApp adoption, minimizing the amount of facts sets:
- the configuration is fetched from 17.1, checking if all required
  variables are available;
- sensitive details retrieved (basically the API key) are stored as
  a secret in the OpenShift cluster;
- a cinder-volume spec is defined using the retrieved configuration
  and the secret, and applied to the control plane.