Add the uni09iota adoption scenario
The scenario follows the configuration of the corresponding
greenfield deployment. The most relevant settings are linked
to the configuration of cinder and glance.

In order to reduce the hardware constraints, and following
the pattern established for the corresponding 18 greenfield job,
there is just a single OpenStack controller.
Moreover, there are additional multipath for cinder/pureFC.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/3399