backend_services: cinder+manila flexible settings
The backend_services test role can set an OpenStackVersion CR with the list
of containers. This is the same mechanism of the update_containers role
of ci-framework, maybe it should use a mechanism similar to the
set_openstack_containers ci-framework role instead for the defaults,
allowing people to set specific containers.

Unfortunately the current implementations requires setting a specific
name for cinder-volume and manila-share containers (volume1 and
share1) respectively. As it was done for ci-framework, the code
needs to be more flexible, and allows users to set:
- the name of all cinder-volume and manila-share containers,
  if they use the standard containers;
- additional full URIs to container images for both cinder-volume
  and manila-share for 3rd-party images.

New variables are needed for this, and their default values match
the previous behavior (a single cinder-volume called volume1,
a single manila-share called share1).

Signed-off-by: Luigi Toscano <ltoscano@redhat.com>