Test Nova After Neutron to avoid reboot timing issue for ping test.
We currently have the reboot of the vm in the nova testing happening
before the neutron ping testing.

But after the reboot, the vm can take a some time to reach the moment
where it reconfigure its network.

This leads to false positive is the ping connectivity test.

Testing the reboot of the vm after the ping test avoid this issue.

Closes: [OSPCIX-1111](https://issues.redhat.com/browse/OSPCIX-1111)